chore: update examples and tests
