+++ coulis/src/core/entities/shorthand.ts
@@ -1,6 +1,5 @@
 import type { PropertiesLike } from "./property";
 
-export type ShortandsLike<Properties extends PropertiesLike> = Record<
-	string,
-	(keyof Properties)[]
->;
+export type ShortandsLike<Properties extends PropertiesLike> =
+	| Record<string, (keyof Properties)[]>
+	| undefined;
