+++ coulis/src/core/entities/state.ts
@@ -1,7 +1,6 @@
-export type StatesLike = Record<
-	string,
-	(input: { className: string; declaration: string }) => string
-> & {
-	// The `base` state cannot be overwritten consumer-side
-	base?: never;
-};
+export type StatesLike =
+	| Record<
+			string,
+			(input: { className: string; declaration: string }) => string
+	  >
+	| undefined;
