+++ coulis/src/core/entities/theme.ts
@@ -1,0 +1,3 @@
+import type { RecordLike } from "./primitive";
+
+export type ThemeLike = RecordLike | undefined;
