+++ examples/vite-ssr-react/src/App.tsx
@@ -1,74 +1,11 @@
 import { useEffect, useState } from "react";
 import type { PropsWithChildren, ReactNode } from "react";
-import { createCoulis, createWebAdapter } from "coulis";
 
-const {
-	createCustomProperties,
+import {
 	createKeyframes,
 	createStyles,
-	createVariants,
 	setGlobalStyles,
-} = createCoulis(createWebAdapter);
-
-const px = (value: number) => `${value}px`;
-
-const tokens = Object.freeze({
-	colors: {
-		black: "black",
-		blue: [
-			"rgb(241,244,248)",
-			"rgb(226,232,240)",
-			"rgb(201,212,227)",
-			"rgb(168,186,211)",
-			"rgb(119,146,185)",
-		],
-		transparent: "transparent",
-		white: "white",
-	},
-	fontSizes: [
-		px(12),
-		px(14),
-		px(16),
-		px(18),
-		px(20),
-		px(22),
-		px(24),
-		px(28),
-		px(30),
-	],
-	fontWeights: ["100", "400", "900"],
-	radii: [px(0), px(4), px(8), px(12), px(999)],
-	spacings: {
-		0: "0px",
-		0.5: "0.5rem",
-		1: "1rem",
-		1.5: "2rem",
-	},
-} as const);
-
-const theme = createCustomProperties({
-	colors: {
-		neutralDark: tokens.colors.black,
-		neutralLight: tokens.colors.white,
-		neutralTransparent: tokens.colors.transparent,
-		surfacePrimary: tokens.colors.blue[4],
-		surfaceSecondary: tokens.colors.blue[2],
-	},
-	fontSizes: {
-		body: tokens.fontSizes[2],
-	},
-	fontWeights: {
-		body: tokens.fontWeights[1],
-	},
-	radii: {
-		full: tokens.radii[4],
-		large: tokens.radii[3],
-		medium: tokens.radii[2],
-		none: tokens.radii[0],
-		small: tokens.radii[1],
-	},
-	spacings: tokens.spacings,
-});
+} from "./helpers/coulis";
 
 setGlobalStyles({
 	"@import":
@@ -81,8 +18,7 @@ setGlobalStyles({
 		boxSizing: "inherit",
 	},
 	".globalClass": {
-		border: "1px solid black",
-		borderRadius: 4,
+		borderRadius: "large",
 	},
 	"html": {
 		boxSizing: "border-box",
@@ -103,74 +39,6 @@ const animationName = createKeyframes({
 	},
 	to: {
 		transform: "scale(1)",
-	},
-});
-
-const styles = createStyles(
-	{
-		accentColor: true,
-		animation: true,
-		backgroundColor: theme.colors,
-		borderRadius: theme.radii,
-		borderStyle: true,
-		color: theme.colors,
-		colorScheme(input: "black" | "white") {
-			return input === "black" ? "dark" : "light";
-		},
-		display: true,
-		flex: true,
-		flexDirection: true,
-		fontSize: theme.fontSizes,
-		fontWeight: theme.fontWeights,
-		gap: true,
-		height: true,
-		margin: theme.spacings,
-		marginBottom: theme.spacings,
-		marginLeft: theme.spacings,
-		marginRight: theme.spacings,
-		marginTop: theme.spacings,
-		padding: theme.spacings,
-		paddingBottom: theme.spacings,
-		paddingLeft: theme.spacings,
-		paddingRight: theme.spacings,
-		paddingTop: theme.spacings,
-		transitionProperty(input: ("background-color" | "color")[]) {
-			return input.join(",");
-		},
-		width: [50, 100],
-	},
-	{
-		shorthands: {
-			marginHorizontal: ["marginLeft", "marginRight"],
-			marginVertical: ["marginTop", "marginBottom"],
-			paddingHorizontal: ["paddingLeft", "paddingRight"],
-			paddingVertical: ["paddingTop", "paddingBottom"],
-		},
-		states: {
-			hover: ({ className, declaration }) =>
-				`${className}:hover{${declaration}}`,
-			large: ({ className, declaration }) =>
-				`@media (min-width: 1024px){${className}{${declaration}}}`,
-			medium: ({ className, declaration }) =>
-				`@media (min-width: 768px){${className}{${declaration}}}`,
-			small: ({ className, declaration }) =>
-				`@media (min-width: 360px){${className}{${declaration}}}`,
-			smallWithHover: ({ className, declaration }) =>
-				`@media (min-width: 360px){${className}:hover{${declaration}}}`,
-		},
-	},
-);
-
-const buttonVariants = createVariants(styles, {
-	color: {
-		accent: { backgroundColor: "surfaceSecondary" },
-		brand: { backgroundColor: "surfacePrimary" },
-		neutral: { backgroundColor: "neutralDark" },
-	},
-	size: {
-		large: { padding: 1.5 },
-		medium: { padding: 1 },
-		small: { padding: 0.5 },
 	},
 });
 
@@ -196,7 +64,7 @@ const App = () => {
 			</Example>
 			<Example title="With static styles">
 				<p
-					className={styles({
+					className={createStyles({
 						backgroundColor: "surfacePrimary",
 						borderRadius: "large",
 						color: {
@@ -221,7 +89,7 @@ const App = () => {
 			</Example>
 			<Example title="With dynamic styles">
 				<p
-					className={styles({
+					className={createStyles({
 						color:
 							counter % 2 === 0
 								? "surfacePrimary"
@@ -233,7 +101,7 @@ const App = () => {
 			</Example>
 			<Example title="With contextual styles">
 				<p
-					className={styles({
+					className={createStyles({
 						backgroundColor: {
 							base: "surfacePrimary",
 							medium: "surfaceSecondary",
@@ -246,7 +114,7 @@ const App = () => {
 			</Example>
 			<Example title="With custom properties">
 				<p
-					className={styles({
+					className={createStyles({
 						borderRadius: "small",
 						borderStyle: "solid",
 					})}
@@ -254,20 +122,9 @@ const App = () => {
 					{TEXT}
 				</p>
 			</Example>
-			<Example title="With variants">
-				<button
-					className={buttonVariants({
-						color: "brand",
-						size: "large",
-					})}
-					type="button"
-				>
-					{TEXT.split(" ")[0]}
-				</button>
-			</Example>
 			<Example title="With keyframes">
 				<div
-					className={styles({
+					className={createStyles({
 						animation: `${animationName} 2000ms linear infinite`,
 						backgroundColor: "surfacePrimary",
 						borderRadius: "medium",
@@ -301,10 +158,12 @@ const Example = ({ title, children }: ExampleProps) => {
 const Layout = ({ children }: PropsWithChildren) => {
 	return (
 		<main
-			className={styles({
+			className={createStyles({
 				display: "flex",
 				flexDirection: "column",
 				gap: 16,
+				paddingHorizontal: 1.5,
+				paddingVertical: 1.5,
 			})}
 		>
 			{children}
